% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVars.R
\name{choose_criteria}
\alias{choose_criteria}
\title{Determine matching criteria}
\usage{
choose_criteria(criteria_list = NULL, k = 200, plot_coverage = TRUE, ...)
}
\arguments{
\item{criteria_list}{list of matching criteria to test in the \code{\link{kpoints}}
function. Each item in the list should be a vector of values (possible matching
criteria) corresponding to each matching variable.}

\item{k}{number of points to use to find solution using \code{\link{kpoints}}
function. Default value is 200.}

\item{plot_coverage}{boolean. Indicates whether the algorithm should display
a barplot of the coverage for each set of criteria. Default is TRUE.}

\item{...}{accepts additional parameters to \code{\link{kpoints}} function.}

\item{matchingvars}{data frame created using \code{\link{makeInputdata}} or
formatted such that: rownames are 'cellnumbers' extracted using the
\code{\link[raster:extract]{raster::extract()}} function, columns 2 and 3
correspond to x and y coordinates, and additional columns correspond to
potential matching variables extracted using the
\code{\link[raster:rasterToPoints]{raster::rasterToPoints()}} function.}

\item{raster_template}{one of the raster layers used for input data.
See \code{\link[raster:area]{raster::area()}}. Note that 'cellnumbers'
column must be present for \code{\link{kpoints}} function to work within this
function.}
}
\value{

}
\description{
Examine summary statistics of matching variables to determine matching criteria
}
\examples{
# Load targetcells data for Target Cells
data(targetcells)
# Create data frame of potential matching variables for Target Cells
allvars <- makeInputdata(targetcells)

# Restrict data to matching variables of interest
matchingvars <- allvars[,c("cellnumbers","x","y","bioclim_01","bioclim_04",
"bioclim_09","bioclim_12","bioclim_15","bioclim_18")]

# Create list of matching criteria to choose:
# Look at 2.5\%, 5\%, & 10\% of range and standard deviation for each variable
range2.5pct <- apply(matchingvars[,4:ncol(matchingvars)],2,
function(x){(max(x)-min(x))*0.025})
range5pct <- apply(matchingvars[,4:ncol(matchingvars)],2,
function(x){(max(x)-min(x))*0.05})
range10pct <- apply(matchingvars[,4:ncol(matchingvars)],2,
function(x){(max(x)-min(x))*0.1})
stddev <- apply(matchingvars[,4:ncol(matchingvars)],2,sd)
criteria_list <- list(range2.5pct, range5pct, range10pct, stddev)

# Compare coverage with various criteria
results2 <- choose_criteria(matchingvars, criteria_list = criteria_list,
n_starts = 1, k = 200, raster_template = targetcells[[1]],
plot_coverage = TRUE)
}
