% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVars.R
\name{criteriaplot}
\alias{criteriaplot}
\title{Create a barplot of results from \code{\link{choose_criteria}} function}
\usage{
criteriaplot(x, criteria_list)
}
\arguments{
\item{x}{a list of results from the \code{\link{choose_criteria}} function.}

\item{criteria_list.}{A list of matching criteria tested in the
\code{\link{choose_criteria}} function.}
}
\value{
a barplot showing the proportion of the study area represented by
different combinations of matching criteria.
}
\description{
Shows a barplot of the proportion of the study area covered by each of the sets
of matching criteria that were tested in the \code{\link{choose_criteria}}
function.
}
\examples{
See \code{\link{choose_criteria}}.
}
