% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendPlot.R
\name{legendPlot}
\alias{legendPlot}
\title{Plot a raster with a custom legend below}
\usage{
legendPlot(
  x = NULL,
  thisVariable = names(x)[1],
  round_dec = 0,
  cols = NULL,
  bks = NULL,
  addpoints = NULL,
  matchingQ = FALSE
)
}
\arguments{
\item{x}{raster. A raster to be plotted.}

\item{thisVariable}{character. The name of the variable in the raster. Defaults
to the name of the raster.}

\item{round_dec}{numeric. The number of decimal places to round the labels
of the legend. Defaults to 0.}

\item{cols}{vector of colors to use in the legend. Default colors range from
yellow to dark brown through 8 distinct colors.}

\item{bks}{vector of breaks to use in designating which values get assigned to
each color in \code{cols}. Must have one more element than \code{cols}. Unless \code{bks} are
designated, this vector will be calculated internally from \code{cols}.}
}
\value{
a plot of the raster with a legend.
}
\description{
Plots a raster with binned colors and adds a legend below the image.
}
\examples{
# Load targetcells data for Target Cells
data(targetcells)
legendPlot(targetcells[[1]])
}
