% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupVars.R
\name{makeInputdata}
\alias{makeInputdata}
\title{Set up input data for Target cells}
\usage{
makeInputdata(x)
}
\arguments{
\item{x}{RasterStack that covers the area of interest and includes all potential
matching variables. Large datasets may be difficult to handle if memory is
limited and it may work better to run the function for several RasterStacks
(removing previous RasterStacks before reading in the next one) and combine
the output into one final data frame. Note: rasters must all have the same
extent, resolution, and crs.}
}
\value{
data frame with a 'cellnumbers' column (these are unique values for each
cell in the rasters used as input and will be used as unique identifiers for
the Target cells in other functions), x and y coordinates, and values for each
of the rasters in the RasterStack. The rownames correspond to the 'cellnumbers'
column.
}
\description{
Prepare data frame of potential matching variables for Target cells from
raster files
}
\examples{
# Load targetcells data for Target Cells
data(targetcells)
# Create data frame of potential matching variables for Target Cells
allvars <- makeInputdata(targetcells)

}
