% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpoints.R
\name{plotcoverage}
\alias{plotcoverage}
\title{Determine optimal number of points (k)}
\usage{
plotcoverage(x)
}
\arguments{
\item{x}{Output from \code{\link{kpoints}} function.}
}
\value{
Plot of the proportion of the study area covered for each value of k,
or if only one value of k was used, reports coverage for that solution.
}
\description{
Determine a number (k) of points that maximize the areal coverage of a study
area using a set of matching variables to determine similarity among sites.
}
\examples{
# Load targetcells data for Target Cells
data(targetcells)

#Create data frame of potential matching variables for Target Cells
allvars <- makeInputdata(targetcells)

# Subset to include only matching variables
matchingvars <- allvars[,c("cellnumbers","x","y","bioclim_01","bioclim_04",
"bioclim_09","bioclim_12","bioclim_15","bioclim_18")]

# Create vector of matching criteria
criteria <- c(0.7,42,3.3,66,5.4,18.4)

# Create sequence of values for k
klist = seq(25,100, by = 25)

# Run kpoints algorithm for klist
# Note: n_starts should be >= 10, it is 1 here to reduce run time.
results3 <- kpoints(matchingvars,criteria = criteria,klist = klist,
n_starts = 1,min_area = 50,iter = 15,
raster_template = targetcells[[1]], verify_stop = FALSE)

# Find optimal number of points (k)
plotcoverage(results3)
}
\author{
Rachel R. Renne
}
