% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMultivariateMatching.R
\docType{package}
\name{rMultivariateMatching}
\alias{rMultivariateMatching}
\title{rMultivariateMatching: A package that uses multivariate matching to generate high-resolution maps}
\description{
The rMultivariateMatching package uses multivariate matching for optimal
selection of sites for simulation and interpolation of simulation output from
spatially sparse simulations to high resolution datasets.
}
\section{}{
NA
}

