% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setsoiltypes}
\alias{setsoiltypes}
\title{Four set soil types used for SOILWAT2 simulations}
\format{
A data frame with 4 rows and 2 variables:

\describe{
\item{sand}{Percentage sand content of soil by weight}
\item{clay}{Percentage clay contet of soil by weight}
}
}
\source{
Soil texture types were determined through an analysis of the soil
texture classes for big sagebrush ecosystems across the western US as described
in Appendix S3 of Palmquist et al. (2021).

\strong{References}\cr
Bradford, J. B., Schlaepfer, D. R., Lauenroth, W. K., Palmquist, K. A.,
Chambers, J. C., Maestas, J. D., & Campbell, S. B. (2019). Climate-driven
shifts in soil temperature and moisture regimes suggest opportunities to
enhance assessments of dryland resilience and resistance. \emph{Front. Ecol. Evol.}
7:358. https://doi.org/10.3389/fevo.2019.00358

Palmquist, K. A., Schlaepfer, D. R., Renne, R. R., Torbit, S. C., Doherty, K. E.,
Remington, T. E., Watson, G., Bradford, J. B. and Lauenroth, W. K., 2021.
Divergent climate change effects on widespread dryland plant communities driven
by climatic and ecohydrological gradients. \emph{Glob. Change Biol.}
https://doi.org/10.1111/gcb.15776

Schlaepfer, D. R., Lauenroth, W. K., & Bradford, J. B. (2012). Ecohydrological
niche of sagebrush ecosystems. \emph{Ecohydrology} 5: 453-466.
https://doi.org/10.1002/eco.238

Schlaepfer, D. R., & Andrews, C. A. (2018). rSFSW2: Simulation Framework for
SOILWAT2. R package version 3.0.0.

Schlaepfer, D. R., & Murphy, R. (2018). rSOILWAT2: An Ecohydrological
Ecosystem-Scale Water Balance Simulation Model. R package version 2.3.2.
}
\usage{
setsoiltypes
}
\description{
A data frame of the percentage of sand and clay by weight of four set soil
types used for SOILWAT2 (Schlaepfer et al., 2012; Schlaepfer & Andrews, 2018;
Schlaepfer & Murphy, 2018) simulations in Bradford et al. (2019). Rownames
provide the treatment number for each soil type (these numbers are appended
as a decimal onto the unique site numbers 'site_id' in the 'subsetcells'
dataset to create the 'site_ids').
}
\keyword{datasets}
