% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendPlot.R
\name{calculateBreaks}
\alias{calculateBreaks}
\title{A helper function for \code{\link{legendPlot}}}
\usage{
calculateBreaks(x, n_cols)
}
\arguments{
\item{x}{raster.}

\item{n_cols}{the number of colors for the legend. Takes input from
\code{\link{legendPlot}}}
}
\value{
a vector of breaks for the legend.
}
\description{
Creates equal sized bins for the legend and display of the raster.
}
\author{
Rachel R. Renne
}
