% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{cverrors}
\alias{cverrors}
\title{Estimate matching errors}
\usage{
cverrors(loocv_output = NULL, first_output_column = NULL)
}
\arguments{
\item{loocv_output}{a data frame produced by the \code{\link{loocv}} function.}

\item{first_output_column}{numeric. The column number of the first column in
the \code{loocv_output} data frame that contains loocv differences in output
variables (should usually be 6 or 8 for one- and two-step matching).}
}
\value{
a named vector where items are the estimated matching error for each
output variable calculated using leave-one-out cross validation of the
Subset cells (simulated sites).
}
\description{
Use the output from \code{\link{loocv}} to estimate matching errors for each
output variable
}
\examples{
########################
# An example where secondarymatch = FALSE
# Load targetcells data for Target Cells
data(targetcells)

# Create data frame of potential matching variables for Target Cells
allvars <- makeInputdata(targetcells)

# Create a mock dataset of output results
output_results <- allvars[rownames(subsetcells),c("cellnumbers","bioclim_02",
                                                  "bioclim_03","bioclim_16",
                                                  "bioclim_17")]

# Create dataset of matchingvars for subsetcells
subset_matchingvars <- matchingvars[rownames(subsetcells),-1]

# Run leave-one-out cross validation of mock output results
loocv_results <- loocv(matchingvars = subset_matchingvars,
                       output_results = output_results,
                       criteria1 = criteria,
                       secondarymatch = FALSE, n_neighbors = 2)

# Calculate estimates of matching error for output variables
estimated_errors <-  cverrors(loocv_output = loocv_results,
first_output_column = 6)
}
\author{
Rachel R. Renne
}
