% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateMatching.R
\name{sd_barplot}
\alias{sd_barplot}
\title{Internal function for \code{\link{evaluateMatching}}}
\usage{
sd_barplot(results)
}
\arguments{
\item{results}{data frame. Output from \code{\link{evaluateMatching}}}
}
\value{
a horizonal barplot of the standard deviation of differences between
target and matched subset cells.
}
\description{
Plots a horizontal barplot of output from \code{\link{evaluateMatching}}
}
\examples{
sd_barplot(results)
}
\author{
Rachel R. Renne
}
